# bpf: optimize map and program metrics collection