… go.sum
… vendor/github.com/cilium/ebpf/info.go
… vendor/github.com/cilium/ebpf/internal/efw/object.go
… vendor/github.com/cilium/ebpf/internal/kallsyms/kallsyms.go
… vendor/github.com/cilium/ebpf/internal/kallsyms/reader.go
… vendor/github.com/cilium/ebpf/map.go
… vendor/github.com/cilium/ebpf/prog.go
… vendor/modules.txt