https://github.com/cilium/ebpf/pull/1791 optimized opening bpf objects and calling obj_info, since it was allocating quite heavily before.

This is the effect on the GetBPFUsage benchmark:

```
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/metrics
cpu: AMD Ryzen 7 3700X 8-Core Processor
               │   old.txt    │              new.txt               │
               │    sec/op    │   sec/op     vs base               │
GetBPFUsage-16   145.43m ± 1%   77.51m ± 1%  -46.70% (p=0.002 n=6)

               │   old.txt    │               new.txt               │
               │     B/op     │     B/op      vs base               │
GetBPFUsage-16   50.38Mi ± 0%   24.24Mi ± 0%  -51.89% (p=0.002 n=6)

               │   old.txt    │              new.txt               │
               │  allocs/op   │  allocs/op   vs base               │
GetBPFUsage-16   358.11k ± 0%   82.67k ± 0%  -76.92% (p=0.002 n=6)
```

This should make the /metrics endpoint even more responsive and reduce the amount of garbage it creates.

@marseel