@marseel 2025-06-04
I can confirm this fixes memory regression we had:
![image](https://github.com/user-attachments/assets/59247349-04c1-4eb6-a7ed-f8b0c3860032)